% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_data}
\alias{add_data}
\title{Add data}
\usage{
add_data(p, data)
}
\arguments{
\item{p}{a datamaps object.}

\item{data}{data.frame.}
}
\description{
Add data
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
                     lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
                     lat = c(51.49999, 40.74998, 39.92889, -33.92001),
                     values = runif(4, 3, 20))

data <- data.frame(name = c("USA", "FRA", "CHN", "RUS", "COG", "DZA", "BRA", "AFG"),
    color = round(runif(8, 1, 10)))

data \%>\%
    datamaps(default = "lightgray") \%>\%
    add_choropleth(name, color) \%>\%
    add_data(coords) \%>\%
    add_bubbles(lon, lat, values, values, city, colors = c("gray", "blue"))

}

