% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{add_graticule}
\alias{add_graticule}
\title{Add graticule}
\usage{
add_graticule(p)
}
\arguments{
\item{p}{a datamaps object.}
}
\description{
Add graticule.
}
\examples{
datamaps(projection = "orthographic") \%>\%
  add_graticule()

}
