% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{add_labels}
\alias{add_labels}
\title{Add labels}
\usage{
add_labels(p)
}
\arguments{
\item{p}{a datamaps object}
}
\description{
Add labels
}
\examples{
states <- data.frame(st = c("AR", "NY", "CA", "IL", "CO", "MT", "TX"),
    val = c(10, 5, 3, 8, 6, 7, 2))

states \%>\%
    datamaps("usa") \%>\%
    add_choropleth(st, val) \%>\%
    add_labels()

}

