% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{add_legend}
\alias{add_legend}
\title{Add legend}
\usage{
add_legend(p)
}
\arguments{
\item{p}{a datamaps object}
}
\description{
Add a legend to the map.
}
\examples{
data <- data.frame(name = c("USA", "FRA", "CHN", "RUS", "COG", "DZA"),
    values = c("N. America", "EU", "Asia", "EU", "Africa", "Africa"))

data \%>\%
    datamaps() \%>\%
    add_choropleth(name, values, colors = c("skyblue", "yellow", "orangered")) \%>\%
    add_legend()

}
