% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{delete_map}
\alias{delete_map}
\title{Remove map}
\usage{
delete_map(proxy)
}
\arguments{
\item{proxy}{a proxy as returned by \code{\link{datamapsProxy}}.}
}
\description{
Remove the map
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  actionButton(
    "delete",
    "Delete map"
  ),
  datamapsOutput("map")
)

server <- function(input, output){
  output$map <- renderDatamaps({
    datamaps()
  })
}

shinyApp(ui, server)
}

}
