% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamaps.R
\name{datamaps}
\alias{datamaps}
\title{Initiate map}
\usage{
datamaps(data, scope = "world", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{scope}{map scope.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{DOM id.}
}
\description{
Initiate a datamaps chart.
}

