% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamaps.R
\name{datamaps}
\alias{datamaps}
\title{Initiate map}
\usage{
datamaps(data, scope = "world", default = "#ABDDA4",
  projection = "equirectangular", responsive = TRUE, width = "100\%",
  height = "100\%", elementId = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{scope}{map scope.}

\item{default}{default color for missing values.}

\item{projection}{map projection.}

\item{responsive}{whether for map to be responsive.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{DOM id.}
}
\description{
Setup a datamaps chart.
}
\examples{
datamaps(projection = "orthographic") \%>\%
  add_graticule()

}
